RES-209 modfied w2iの仕様変更
戻り値の型をlong longに変更
未知語の場合-1を返す

+++ W2V.hpp
@@ -18,7 +18,7 @@ namespace W2V {
 		Matrix_();
         virtual ~Matrix_();
 		void read(const std::string& file);
-		int w2i(const std::string& w);
+		long long w2i(const std::string& w);
 		const vector& i2v(int i);
 		long long getNumWords() { return words; }
 		long long getSize() { return size; }
